+++ packages/grpc/CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Changed
 
-- Bump dependency on [@bufbuild/protobuf] to 2.6.1 ([#1200](https://github.com/cerbos/cerbos-sdk-javascript/pull/1200))
+- Bump dependency on [@bufbuild/protobuf] to 2.6.2 ([#1200](https://github.com/cerbos/cerbos-sdk-javascript/pull/1200), [#1202](https://github.com/cerbos/cerbos-sdk-javascript/pull/1202))
 
 ### Removed